Ticket#11138, CP-1058 - CRM - Ajuste personFilter cuentas
