--- custom/clients/base/api/GetPersonFilterApi.php
@@ -1,96 +1,0 @@
-<?php
-if (!defined('sugarEntry') || !sugarEntry) die('Not A Valid Entry Point');
-require_once("clients/base/api/PersonFilterApi.php");
-
-class GetPersonFilterApi extends FilterApi {
-    
-    public function registerApiRest() 
-    {
-        return array(
-            'UserSearch' => array(
-                'reqType' => 'GET',
-                'path' => array('Users'),
-                'jsonParams' => array('filter'),
-                'pathVars' => array('module_list'),
-                'method' => 'filterList',
-                'shortHelp' => 'Search User records',
-                'longHelp' => 'include/api/help/module_filter_get_help.html',
-                'exceptions' => array(
-                    // Thrown in filterList and filterListSetup
-                    'SugarApiExceptionInvalidParameter',
-                    // Thrown in filterListSetup and parseArguments
-                    'SugarApiExceptionNotAuthorized',
-                    'SugarApiExceptionError',
-                )
-            ),
-            'EmployeeSearch' => array(
-                'reqType' => 'GET',
-                'path' => array('Employees'),
-                'jsonParams' => array('filter'),
-                'pathVars' => array('module_list'),
-                'method' => 'filterList',
-                'shortHelp' => 'Search Employee records',
-                'longHelp' => 'include/api/help/module_filter_get_help.html',
-                'exceptions' => array(
-                    // Thrown in filterList and filterListSetup
-                    'SugarApiExceptionInvalidParameter',
-                    // Thrown in filterListSetup and parseArguments
-                    'SugarApiExceptionNotAuthorized',
-                    'SugarApiExceptionError',
-                )
-            ),
-        );
-    }
-
-    
-    public function filterList(ServiceBase $api, array $args, $acl = 'list')
-    {
-        if (!empty($args['q'])) {
-            return $this->globalSearch($api, $args);
-        }
-
-        $args['module'] = $args['module_list'];
-
-        $api->action = 'list';
-        list($args, $q, $options, $seed) = $this->filterListSetup($api, $args);
-
-        $this->getCustomWhereForModule($args['module_list'], $q);
-
-        return $this->runQuery($api, $args, $q, $options, $seed);
-    }
-
-
-    public function globalSearch(ServiceBase $api, array $args) {
-        $api->action = 'list';
-        // This is required to keep the loadFromRow() function in the bean from making our day harder than it already is.
-        $GLOBALS['disable_date_format'] = true;
-        $search = new UnifiedSearchApi();
-        $options = $search->parseSearchOptions($api,$args);
-        $options['custom_where'] = $this->getCustomWhereForModule($args['module_list']);
-
-        $searchEngine = new SugarSpot();
-        $options['resortResults'] = true;
-        $recordSet = $search->globalSearchSpot($api,$args,$searchEngine,$options);
-        
-        return $recordSet;
-    }
-
-    
-    protected function getCustomWhereForModule($module, $query = null) {
-
-        if ($query instanceof SugarQuery) {
-            if ($module == 'Employees') {
-                $query->where()->equals('employee_status', 'Active')->equals('show_on_employees','1');
-                return;
-            }
-            $query->where()->equals('portal_only', '0');
-            return;
-        }
-
-        if ($module == 'Employees') {
-            return "users.employee_status = 'Active' AND users.show_on_employees = 1";
-        }
-        
-        return "users.portal_only = 0";
-    }
-}
